requirements.txt
alabaster==0.7.12
appnope==0.1.2
Babel==2.9.1
backcall==0.2.0
boto3==1.20.24
botocore==1.23.24
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
dnspython==2.1.0
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
fett==0.3.2
idna==3.3
imagesize==1.3.0
ipykernel==6.5.0
ipython==7.29.0
jedi==0.18.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.9.1
kiwisolver==1.3.2
livereload==2.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
networkx==2.6.3
numpy==1.21.2
packaging==21.3
pandas==1.2.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pipe==1.6.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
Pygments==2.10.0
pymongo==3.12.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jsonrpc-server==0.3.4
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
rstcheck==3.3.1
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
snooty-lextudio==1.11.5
snowballstemmer==2.2.0
Sphinx==4.3.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.0
ujson==1.35
urllib3==1.26.7
watchdog==1.0.2
wcwidth==0.2.5
